FIX simplifying logic - try to flatten render logic via removing extra MeshType indirection   * copy flags to node - make MeshType findable via lua side   * use SID for identity

+++ src/ki/size.h
@@ -8,7 +8,7 @@ namespace ki
 
     typedef uint32_t node_id;
     typedef uint16_t mesh_id;
-    typedef uint16_t type_id;
+    typedef uint32_t type_id;
 
     typedef uint32_t material_id;
     typedef int32_t material_index;
